backend/uclapi/requirements.txt
autopep8==1.4.4
amqp==2.5.2
boto3==1.10.13
botocore==1.13.13
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
ciso8601==1.0.8
cx-Oracle==6.4.1
decorator==4.3.0
deepdiff==3.3.0
Django==2.1.1.
django-cors-headers==2.4.0
django-dotenv==1.4.2
django-mock-queries==2.1.3
django-storages==1.7.2
djangorestframework==3.10.3
django-postgrespool2==1.0.0
docutils==0.15.2
eventlet==0.25.1
gevent==1.4.0
freezegun==0.3.12
flake8==3.7.9
greenlet==0.4.15
gunicorn==19.9.0
hiredis==1.0.0
idna==2.8
jinjasql==0.1.7
jmespath==0.9.4
jsonpickle==1.2
keen==0.5.1
kombu==4.6.3
lxml==4.4.1
Markupsafe==1.1.1
mock==3.0.5
model-mommy==2.0.0
pbr==5.4.3
pycryptodome==3.9.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
raven==6.10.0
redis==3.3.11
requests==2.22.0
requests-file==1.4.3
requests-futures==1.0.0
setproctitle==1.1.10
six==1.13.0
tldextract==2.2.2
tqdm==4.37.0
urllib3==1.25.6
validators==0.14.0
vine==1.3.0
